LayoutTests/svg/animations/svgpath-animation-1-expected.txt

SVG 1.1 dynamic animation tests

Test fallback to 'to' animation if user specifies 'by' animation on path. You should see a green 100x100 path and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS path.pathSegList.getItem(0).x is 40
PASS path.pathSegList.getItem(0).y is 40
PASS path.pathSegList.getItem(0).x is 20
PASS path.pathSegList.getItem(0).y is 20
PASS path.pathSegList.getItem(0).x is 0
PASS path.pathSegList.getItem(0).y is 0
PASS path.pathSegList.getItem(0).x is 40
PASS path.pathSegList.getItem(0).y is 40
PASS successfullyParsed is true

TEST COMPLETE
